Change livedata to stateflow in viewmodels whenever possible

+++ app/src/main/java/es/upsa/mimo/gamercollection/presentation/games/GamesViewModel.kt
@@ -5,8 +5,6 @@ import android.content.res.Resources
 import android.view.Gravity
 import android.widget.LinearLayout
 import android.widget.NumberPicker
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
@@ -21,6 +19,8 @@ import es.upsa.mimo.gamercollection.extensions.getPickerParams
 import es.upsa.mimo.gamercollection.extensions.setup
 import es.upsa.mimo.gamercollection.utils.ScrollPosition
 import javax.inject.Inject
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 @HiltViewModel
@@ -29,15 +29,15 @@ class GamesViewModel @Inject constructor(
 ) : ViewModel() {
 
     //region Private properties
-    private val _gamesLoading = MutableLiveData<Boolean>()
-    private val _gamesError = MutableLiveData<ErrorModel>()
-    private val originalGames = MutableLiveData<List<Game>>()
-    private val _games = MutableLiveData<List<Game>>()
-    private val _gamesCount = MutableLiveData<List<Game>>()
-    private val _gameDeleted = MutableLiveData<Int?>()
-    private var _state = MutableLiveData<String?>(null)
-    private var _filters = MutableLiveData<FilterModel?>(null)
-    private var _scrollPosition = MutableLiveData(ScrollPosition.TOP)
+    private val _gamesLoading = MutableStateFlow<Boolean?>(null)
+    private val _gamesError = MutableStateFlow<ErrorModel?>(null)
+    private val _games = MutableStateFlow<List<Game>>(emptyList())
+    private val _gamesCount = MutableStateFlow<List<Game>>(emptyList())
+    private val _gameDeleted = MutableStateFlow<Int?>(null)
+    private var _state = MutableStateFlow<String?>(null)
+    private var _filters = MutableStateFlow<FilterModel?>(null)
+    private var _scrollPosition = MutableStateFlow(ScrollPosition.TOP)
+    private val originalGames = MutableStateFlow<List<Game>>(emptyList())
     private var sortParam: String = SharedPreferencesHelper.sortParam
     private var isSortOrderAscending = SharedPreferencesHelper.isSortOrderAscending
     private var query: String? = null
@@ -52,14 +52,14 @@ class GamesViewModel @Inject constructor(
         get() = SharedPreferencesHelper.filterDateFormat
     val swipeRefresh: Boolean
         get() = SharedPreferencesHelper.swipeRefresh
-    val gamesLoading: LiveData<Boolean> = _gamesLoading
-    val gamesError: LiveData<ErrorModel> = _gamesError
-    val games: LiveData<List<Game>> = _games
-    val gamesCount: LiveData<List<Game>> = _gamesCount
-    val gameDeleted: LiveData<Int?> = _gameDeleted
-    val state: LiveData<String?> = _state
-    val filters: LiveData<FilterModel?> = _filters
-    val scrollPosition: LiveData<ScrollPosition> = _scrollPosition
+    val gamesLoading: StateFlow<Boolean?> = _gamesLoading
+    val gamesError: StateFlow<ErrorModel?> = _gamesError
+    val games: StateFlow<List<Game>> = _games
+    val gamesCount: StateFlow<List<Game>> = _gamesCount
+    val gameDeleted: StateFlow<Int?> = _gameDeleted
+    val state: StateFlow<String?> = _state
+    val filters: StateFlow<FilterModel?> = _filters
+    val scrollPosition: StateFlow<ScrollPosition> = _scrollPosition
     //endregion
 
     //region Public methods
@@ -81,9 +81,7 @@ class GamesViewModel @Inject constructor(
             originalGames.value = games
 
             if (!_state.value.isNullOrBlank()) {
-                _games.value = originalGames.value?.filter { game ->
-                    game.state == _state.value
-                } ?: listOf()
+                _games.value = games.filter { it.state == _state.value }
             } else {
                 _games.value = games
             }
@@ -140,17 +138,17 @@ class GamesViewModel @Inject constructor(
 
     fun deleteGame(position: Int) {
         viewModelScope.launch {
-            _games.value?.get(position)?.let { game ->
+            if (_games.value.isEmpty()) return@launch
 
-                _gamesLoading.value = true
-                gameRepository.deleteGame(game)
-                _games.value?.first { it.id == game.id }?.let { removed ->
-                    _games.value = _games.value?.minus(removed)
-                }
-                _gameDeleted.value = position
-                _gameDeleted.value = null
-                _gamesLoading.value = false
+            _gamesLoading.value = true
+            val game = _games.value[position]
+            gameRepository.deleteGame(game)
+            _games.value.firstOrNull { it.id == game.id }?.let {
+                _games.value = _games.value.minus(it)
             }
+            _gameDeleted.value = position
+            _gameDeleted.value = null
+            _gamesLoading.value = false
         }
     }
 
@@ -162,9 +160,7 @@ class GamesViewModel @Inject constructor(
     fun setState(newState: String?) {
         _state.value = newState
         if (!newState.isNullOrBlank()) {
-            _games.value = originalGames.value?.filter { game ->
-                game.state == newState
-            } ?: listOf()
+            _games.value = originalGames.value.filter { it.state == newState }
         } else {
             _games.value = originalGames.value
         }
